+++ src/datetime.ts
@@ -1,6 +1,5 @@
 import { modulo, Nomalize } from "emnorst";
 import { IDuration } from "./duration";
-import { Interval } from "./interval";
 import { dateToString, timeToString } from "./string";
 
 export interface IDate {
@@ -102,9 +101,6 @@ const dateTimeDefaults: IDateTime = {
 };
 
 export class DateTime implements IDateTime {
-    static interval(start: DateTimeable, end: DateTimeable): Interval {
-        return new Interval(DateTime.from(start), DateTime.from(end));
-    }
     static now(): DateTime {
         return DateTime.from(Date.now());
     }